# docs/swagger.yaml
definitions:
  models.Category:
    properties:
      created_at:
        type: string
      description:
        type: string
      display_order:
        type: integer
      id:
        type: integer
      image_url:
        type: string
      name:
        type: string
      updated_at:
        type: string
    type: object
  models.OptionValues:
    properties:
      id:
        type: integer
      option_id:
        type: integer
      option_price:
        type: number
      value:
        type: string
    type: object
...
info:
  contact: {}
  description: Super Ordening System
  title: WacDo
  version: "1.0"
paths:
  /categories:
    get:
      description: Retrieve a list of all product categories
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Category'
            type: array
        "500":
          description: Internal error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get all categories
      tags:
      - Categories
    post:
      consumes:
      - application/json
      description: Create a new product category with the provided details
      parameters:
      - description: Category details
        in: body
        name: category
        required: true
        schema:
          $ref: '#/definitions/models.Category'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Category'
        "400":
          description: Invalid data or category already exists
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create a new category
      tags:
      - Categories
  /categories/{id}:
    delete:
      description: Delete a category by ID (only if not in use by any products)
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Category deleted
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid ID
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Category not found
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Category still in use
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete a category
      tags:
      - Categories
    get:
      description: Retrieve a single category by its ID
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Category'
        "400":
          description: Invalid ID
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Category not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get a category by ID
      tags:
      - Categories
    put:
      consumes:
      - application/json
      description: Update an existing category by ID
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated category details
        in: body
        name: category
        required: true
        schema:
          $ref: '#/definitions/models.Category'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Category'
        "400":
          description: Invalid ID or data
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Category not found
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Category name already exists
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update a category
      tags:
      - Categories
...
  /options/{option_id}/values:
    get:
      description: Retrieve all values belonging to a specific option
      parameters:
      - description: Option ID
        in: path
        name: option_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.OptionValues'
            type: array
        "400":
          description: Invalid option ID
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Option not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get values by option
      tags:
      - Option Values
    post:
      consumes:
      - application/json
      description: Create one or more values for a product option (e.g. "Small", "Medium",
        "Large" for Size)
      parameters:
      - description: Option ID
        in: path
        name: option_id
        required: true
        type: integer
      - description: Array of option value details
        in: body
        name: values
        required: true
        schema:
          items:
            $ref: '#/definitions/models.OptionValues'
          type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.OptionValues'
            type: array
        "400":
          description: Invalid data or option not found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create option values
      tags:
      - Option Values
...
  /options/values/{id}:
    delete:
      description: Delete an option value by ID
      parameters:
      - description: Option Value ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Option value deleted
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid ID
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Option value not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete an option value
      tags:
      - Option Values
    get:
      description: Retrieve a single option value by its ID
      parameters:
      - description: Option Value ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.OptionValues'
        "400":
          description: Invalid ID
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Option value not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get an option value by ID
      tags:
      - Option Values
    put:
      consumes:
      - application/json
      description: Update an existing option value by ID
      parameters:
      - description: Option Value ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated option value details
        in: body
        name: value
        required: true
        schema:
          $ref: '#/definitions/models.OptionValues'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.OptionValues'
        "400":
          description: Invalid ID or data
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Option value not found
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Value already exists for this option
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update an option value
      tags:
      - Option Values
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"